+++ backend/.env
@@ -1,0 +1,1 @@
+MONGO_DB_PASSWORD: qeGhPVPgUJcPrktp
