+++ backend/package.json
@@ -4,9 +4,12 @@
 	"description": "",
 	"main": "main.js",
 	"scripts": {
-		"start": "exec node main.js"
+		"start": "exec node main.js",
+		"start dev": "node src/main.js"
 	},
 	"dependencies": {
+		"@types/react": "^19.1.8",
+		"@types/react-dom": "^19.1.6",
 		"body-parser": "^1.19.0",
 		"cors": "^2.8.5",
 		"express": "^4.17.1",
