+++ backend/src/routes/users.routes.js
@@ -9,13 +9,8 @@ router.get('/health', (req, res) => {
     res.status(200).send('OK');
 });
 
-// Root endpoint for health check
-router.get('/', (req, res) => {
-    res.status(200).send('Server is running');
-});
-
 // Endpoint to create a user
-router.post('/user/create', async (req, res) => {
+router.post('/users', async (req, res) => {
     try {
         const newUser = new User(req.body);
         const user = await newUser.save();
@@ -26,7 +21,7 @@ router.post('/user/create', async (req, res) => {
 });
 
 // Endpoint to get a user
-router.get('/user/:username', async (req, res) => {
+router.get('/users/:username', async (req, res) => {
     try {
         const user = await User.findOne({ username: req.params.username });
         if (!user) 
@@ -38,9 +33,9 @@ router.get('/user/:username', async (req, res) => {
 });
 
 // Endpoint to add a friend
-router.post('/user/addFriend', async (req, res) => {
+router.post('/users/:username/friend/:friendUsername', async (req, res) => {
     try{
-        const { userUsername, friendUsername } = req.body;
+        const { userUsername, friendUsername } = req.params;
         const user = await User.findOne( { username: userUsername } );
         const friend = await User.findOne({username: friendUsername});
 
@@ -62,7 +57,7 @@ router.post('/user/addFriend', async (req, res) => {
 });
 
 // Endpoint to update challenge
-router.put('/user/updateChallenge', async (req, res) => {
+router.put('/users/:username/challenge/:challengeId', async (req, res) => {
     try {
         const { username, challenge } = req.body;
         const { challengeId, completed, answer, doc, completedTime } = challenge;
@@ -115,7 +110,7 @@ router.put('/user/updateChallenge', async (req, res) => {
 });
 
 // Endpoint to get specific challenge for user
-router.get('/user/:username/challenge/:challengeId', async (req, res) => {
+router.get('/users/:username/challenge/:challengeId', async (req, res) => {
     try {
         const { username, challengeId } = req.params;
         const user = await User.findOne({ username: username });
@@ -125,39 +120,6 @@ router.get('/user/:username/challenge/:challengeId', async (req, res) => {
         const userChallengeId = user.challenges.find((uc) => uc.challengeId === challengeId);
         const challenge = UserChallenge.findOne({_id: userChallengeId})
         res.status(200).json(challenge);
-    } catch (err) {
-        res.status(500).json({ error: err.message });
-    }
-});
-
-// Endpoint to get specific (today's) challenge
-router.get('/challenge', async (req, res) => {
-    try {
-        const challenge = await Challenge.findOne().sort({ createdAt: -1 });
-        if (!challenge) 
-            return res.status(404).json({ error: 'Challenge not found' });
-
-        const response = {
-            title: challenge.title,
-            description: challenge.description,
-            pointsReward: challenge.pointsReward,
-            isActive: challenge.isActive,
-            createdAt: challenge.createdAt,
-            endDate: challenge.endDate,
-            challengeId: challenge._id
-        }
-        res.status(200).json(response);
-    } catch (err) {
-        res.status(500).json({ error: err.message });
-    }
-});
-
-// Endpoint to get all challenges
-router.get('/challenges', async (req, res) => {
-    try {
-        const challenges = await Challenge.find({}, { title: 1, description: 1, pointsReward: 1,
-                                                        isActive: 1, createdAt: 1, endDate: 1, _id: 1 });
-        res.status(200).json(challenges);
     } catch (err) {
         res.status(500).json({ error: err.message });
     }
